requirements.txt
fastapi~=0.109.0
PyJWT~=2.8.0
starlette~=0.35.1
pydantic~=2.5.3
SQLAlchemy~=2.0.25
aiosqlite~=0.19.0
python-dotenv~=1.0.0
bcrypt~=4.1.2
uvicorn~=0.26.0
Faker~=22.5.1
aiohttp~=3.9.1
yarl~=1.9.4